"refactor: 엔티티 이름 수정"
